+++ test/action/NewsActions.spec.js
@@ -1,4 +1,4 @@
-var jsdom = require('mocha-jsdom')
+import jsdom from 'mocha-jsdom'
 import React from 'react';
 import chai from 'chai';
 import { expect, assert } from 'chai';
@@ -6,7 +6,7 @@ import sinon from 'sinon';
 import {mount, render, shallow} from 'enzyme'
 import chaiEnzyme from 'chai-enzyme'
 import { EventEmitter } from 'events'
-import NewsActions from '../../src/actions/NewsActions';
+import NewsActions from './../../src/actions/NewsActions';
 chai.use(chaiEnzyme())
 const root = new EventEmitter()
 
@@ -19,26 +19,26 @@ describe('News Actions', () => {
 
 
   it('should exist', () => {
-    expect(NewsAction).to.exist;
+    expect(NewsActions).to.exist;
   });
 
   it('getHeadlines should exist', () => {
-    expect(NewsAction.getHeadlines).to.exist;
+    expect(NewsActions.getHeadlines).to.exist;
   });
 
 
   it('getHeadlines should be a functon', () => {
-    expect(NewsAction.getHeadlines).to.be.a('function');
+    expect(NewsActions.getHeadlines).to.be.a('function');
   });
 
   it('getSources should be a functon', () => {
-      NewsAction.getSources()
-    expect(NewsAction.getSources).to.be.a('function');
+      NewsActions.getSources()
+    expect(NewsActions.getSources).to.be.a('function');
   });
 
   it('getSources should be a functon', () => {
-   NewsAction.getHeadlines()
-    expect(NewsAction.getSources).to.be.a('function');
+   NewsActions.getHeadlines()
+    expect(NewsActions.getSources).to.be.a('function');
   });
 })
