+++ test/api/NewsApi.spec.js
@@ -3,7 +3,7 @@ import { expect, assert } from 'chai';
 import sinon from 'sinon';
 import axios from 'axios';
 import dispatcher from '../../src/dispatcher/dispatcher';
-import NewsApi from '../../src/utils/NewsApi';
+import NewsApi from './../../src/utils/NewsApi';
 
 
 describe('News Api', () => {
