+++ test/components/Home.spec.js
@@ -1,6 +1,6 @@
 var jsdom = require('mocha-jsdom')
 import React from 'react';
-import Home from '../../src/components/home.js';
+import Home from './../../src/components/home.js';
 import { expect } from 'chai';
 import { shallow } from 'enzyme';
 import sinon from 'sinon';
